data collection improvements 4 (#88)
* improved features

* new functions and code fixes

* code fixes and improvements

+++ src/modules/Common/public/Get-SdnDiagnosticLog.ps1
@@ -26,28 +26,31 @@ function Get-SdnDiagnosticLog {
 
     try {
         $localLogDir = "C:\Windows\tracing\SDNDiagnostics\Logs"
+        [System.IO.FileInfo]$OutputDirectory = Join-Path -Path $OutputDirectory.FullName -ChildPath "SdnDiagnosticLogs"
 
-        if (!(Test-Path -Path $localLogDir)) {
-            "No SdnDiagnostics folder found, this need to run on SDN Infrastructure Nodes" | Trace-Output -Level:Warning
-            return
+        "Collect SdnDiagnostics logs between {0} and {1} UTC" -f $FromDate.ToUniversalTime(), (Get-Date).ToUniversalTime() | Trace-Output
+
+        if (!(Initialize-DataCollection -FilePath $OutputDirectory.FullName -MinimumGB 5)) {
+            throw New-Object System.Exception("Unable to initialize environment for data collection")
         }
 
-        "Collect SdnDiagnostics logs between {0} and {1}" -f $FromDate, (Get-Date) | Trace-Output -Verbose
+        $logFiles = Get-ChildItem -Path $localLogDir | Where-Object { $_.LastWriteTime -ge $FromDate }
 
-        # Create local directory for SdnDiagnostics logs
-        [System.IO.FileInfo]$OutputDirectory = Join-Path -Path $OutputDirectory.FullName -ChildPath "SdnDiagnostics"
-        if (!(Test-Path -Path $OutputDirectory.FullName -PathType Container)) {
-            $null = New-Item -Path $OutputDirectory.FullName -ItemType Directory -Force
+        # add a failsafe to ensure that we have enough disk space with some extra overhead
+        [System.Char]$driveLetter = (Split-Path -Path $OutputDirectory.FullName -Qualifier).Replace(':','')
+        $minimumDiskSpace = [float](Get-FolderSize -FileName $logFiles.FullName -Total).GB * 1.5
+        "Validating that {0} GB exists for operation" -f $minimumDiskSpace | Trace-Output -Level:Verbose
+        if (-NOT (Confirm-DiskSpace -DriveLetter $driveLetter -MinimumGB $minimumDiskSpace)){
+            throw New-Object System.Exception("Insufficient disk space to perform operation. Reduce the amount of hours to collect logs or free up disk space.")
         }
 
-        $sdnDiagLogs = Get-ChildItem -Path $localLogDir | Where-Object { $_.LastWriteTime -ge $FromDate }
-        foreach ($sdnDiagLog in $sdnDiagLogs) {
-            Copy-Item $sdnDiagLog.FullName -Destination $OutputDirectory.FullName
-        }
+        # copy the log files from the default log directory to the output directory
+        "Copying {0} files to {1}" -f $logFiles.Count, $OutputDirectory.FullName | Trace-Output -Level:Verbose
+        Copy-Item -Path $logFiles.FullName -Destination $OutputDirectory.FullName -Force
 
         # once we have copied the files to the new location we want to compress them to reduce disk space
         # if confirmed we have a .zip file, then remove the staging folder
-        Compress-Archive -Path "$($OutputDirectory.FullName)\*" -Destination $OutputDirectory.FullName -CompressionLevel Optimal
+        Compress-Archive -Path "$($OutputDirectory.FullName)\*" -Destination $OutputDirectory.FullName -CompressionLevel Optimal -Force
         if (Test-Path -Path "$($OutputDirectory.FullName).zip" -PathType Leaf) {
             Remove-Item -Path $OutputDirectory.FullName -Force -Recurse
         }
